services/api/requirements.txt
httpx==0.27.2
fastapi
uvicorn[standard]
httpx==0.27.2
python-json-logger==2.0.7
opentelemetry-exporter-otlp-proto-http==1.27.0
ruff==0.4.6
black==24.4.2
mypy==1.10.0
pip-audit==2.7.2
bandit==1.7.7